+++ src/lib.rs
@@ -104,6 +104,7 @@ pub use crate::algorithms::connectivity::transitive_closure::transitive_closure;
 pub use crate::algorithms::constructors::kary_tree::{TreeMode, kary_tree};
 pub use crate::algorithms::constructors::ring::{cycle_graph, path_graph, ring_graph};
 pub use crate::algorithms::constructors::star::{StarMode, star_graph};
+pub use crate::algorithms::constructors::symmetric_tree::symmetric_tree;
 pub use crate::algorithms::constructors::wheel::{WheelMode, wheel_graph};
 pub use crate::algorithms::games::barabasi::barabasi_game_bag;
 pub use crate::algorithms::games::barabasi_aging::barabasi_aging_game;